Added benchmark test
Since the `main` branch does not have the MuxReplies for JetStream options, I ran the
bench on main without the `MuxReplies=true` option and duplicated the run without
the muxer so that it can be compared with the PR.

This is the result on one computer (Intel based):

```
        [ === BenchJetStreamPubAsync === ]

                                      Name   Base msgs/sec   Bench msgs/sec   Difference

   Ctxs=1/MaxPending=1000/MuxReplies=false          478379           489425       +2.31%
    Ctxs=1/MaxPending=1000/MuxReplies=true          478379           485488       +1.49%
   Ctxs=1/MaxPending=5000/MuxReplies=false          482236           489199       +1.44%
    Ctxs=1/MaxPending=5000/MuxReplies=true          482236           502488       +4.20%
   Ctxs=2/MaxPending=1000/MuxReplies=false          533308           533376       +0.01%
    Ctxs=2/MaxPending=1000/MuxReplies=true          533308           522602       -2.01%
   Ctxs=2/MaxPending=5000/MuxReplies=false          514850           513382       -0.29%
    Ctxs=2/MaxPending=5000/MuxReplies=true          514850           499895       -2.90%

             [ === BenchRequestReply === ]

                                      Name   Base msgs/sec   Bench msgs/sec   Difference

                                  1 thread           11648            12522       +7.50%
                                10 threads           57465            67551      +17.55%
                               100 threads           44782            51964      +16.04%
```

However, on another computer that is ARM64 based, I see a degradation for the req/reply
with 100 threads:

```
        [ === BenchJetStreamPubAsync === ]

                                      Name   Base msgs/sec   Bench msgs/sec   Difference

   Ctxs=1/MaxPending=1000/MuxReplies=false          650696           662329       +1.79%
    Ctxs=1/MaxPending=1000/MuxReplies=true          650696           646994       -0.57%
   Ctxs=1/MaxPending=5000/MuxReplies=false          641985           638520       -0.54%
    Ctxs=1/MaxPending=5000/MuxReplies=true          641985           625851       -2.51%
   Ctxs=2/MaxPending=1000/MuxReplies=false          624024           630602       +1.05%
    Ctxs=2/MaxPending=1000/MuxReplies=true          624024           620501       -0.56%
   Ctxs=2/MaxPending=5000/MuxReplies=false          572678           573279       +0.10%
    Ctxs=2/MaxPending=5000/MuxReplies=true          572678           568538       -0.72%

             [ === BenchRequestReply === ]

                                      Name   Base msgs/sec   Bench msgs/sec   Difference

                                  1 thread           16137            16794       +4.07%
                                10 threads           53078            58417      +10.06%
                               100 threads           42048            38273       -8.98%
```

I have experimented by removing the muxer's lock and use the connection lock and it
seems to resolve the performance degradation. Will make the change in different
commit and post the result of re-running the benchmark.
