+++ test/test.c
@@ -16,31 +16,24 @@
 #include <stdio.h>
 #include <string.h>
 #include <errno.h>
-#ifdef _WIN32
-#else
-#include <dirent.h>
-#include <sys/stat.h>
-#include <execinfo.h>
-#endif
 
-#include "buf.h"
-#include "timer.h"
-#include "url.h"
-#include "opts.h"
+#include "../src/buf.h"
+#include "../src/timer.h"
+#include "../src/url.h"
+#include "../src/opts.h"
 #include "../src/util.h"
-#include "hash.h"
-#include "conn.h"
-#include "sub.h"
-#include "msg.h"
-#include "stats.h"
-#include "crypto.h"
-#include "nkeys.h"
-#include "parser.h"
-#include "js.h"
-#include "kv.h"
-#include "microp.h"
-#include "glib/glibp.h"
-#include "object.h"
+#include "../src/hash.h"
+#include "../src/conn.h"
+#include "../src/sub.h"
+#include "../src/msg.h"
+#include "../src/crypto.h"
+#include "../src/nkeys.h"
+#include "../src/parser.h"
+#include "../src/js.h"
+#include "../src/kv.h"
+#include "../src/microp.h"
+#include "../src/glib/glibp.h"
+#include "../src/object.h"
 
 #if defined(NATS_HAS_STREAMING)
 
@@ -86,17 +79,6 @@ static const char *natsStreamingServerExe = "nats-streaming-server";
 
 natsMutex *slMu  = NULL;
 natsHash  *slMap = NULL;
-
-#define test(s)         { printf("#%02d ", ++tests); printf("%s", (s)); fflush(stdout); }
-#define testf(s, ...)   { printf("#%02d ", ++tests); printf((s), __VA_ARGS__); fflush(stdout); }
-
-#ifdef _WIN32
-#define testCond(c)         if(c) { printf("PASSED\n"); fflush(stdout); } else { printf("FAILED\n"); nats_PrintLastErrorStack(stdout); fflush(stdout); failed=true; return; }
-#define testCondNoReturn(c) if(c) { printf("PASSED\n"); fflush(stdout); } else { printf("FAILED\n"); nats_PrintLastErrorStack(stdout); fflush(stdout); failed=true; }
-#else
-#define testCond(c)         if(c) { printf("\033[0;32mPASSED\033[0;0m\n"); fflush(stdout); } else { printf("\033[0;31mFAILED\033[0;0m\n"); nats_PrintLastErrorStack(stdout); fflush(stdout); failed=true; return; }
-#define testCondNoReturn(c) if(c) { printf("\033[0;32mPASSED\033[0;0m\n"); fflush(stdout); } else { printf("\033[0;31mFAILED\033[0;0m\n"); nats_PrintLastErrorStack(stdout); fflush(stdout); failed=true; }
-#endif
 
 static const char *testServers[] = {"nats://127.0.0.1:1222",
                                     "nats://127.0.0.1:1223",
@@ -23796,99 +23778,6 @@ void test_SSLURLSchemeNonTLSBuild(void)
 }
 
 static void
-rmtree(const char *path)
-{
-#ifdef _WIN32
-    WIN32_FIND_DATA ffd;
-    HANDLE          hFind = INVALID_HANDLE_VALUE;
-    char            *dir  = NULL;
-
-    if (nats_asprintf(&dir, "%s\\*", path) < 0)
-        abort();
-
-    hFind = FindFirstFile(dir, &ffd);
-    if (hFind == INVALID_HANDLE_VALUE)
-    {
-        free(dir);
-        return;
-    }
-
-    do
-    {
-        char *fullPath = NULL;
-
-        if (!strcmp(ffd.cFileName, ".") || !strcmp(ffd.cFileName, ".."))
-            continue;
-
-        if (nats_asprintf(&fullPath, "%s\\%s", path, ffd.cFileName) < 0)
-            abort();
-
-        if (ffd.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY)
-            rmtree(fullPath);
-        else
-            DeleteFile(fullPath);
-        free(fullPath);
-    }
-    while (FindNextFile(hFind, &ffd) != 0);
-
-    FindClose(hFind);
-    RemoveDirectory(path);
-    free(dir);
-
-#else
-    DIR             *dir;
-    struct stat     statPath, statEntry;
-    struct dirent   *entry;
-
-    memset(&statPath, 0, sizeof(struct stat));
-
-    stat(path, &statPath);
-    if (S_ISDIR(statPath.st_mode) == 0)
-        return;
-
-    if ((dir = opendir(path)) == NULL)
-        return;
-
-    while ((entry = readdir(dir)) != NULL)
-    {
-        char *fullPath = NULL;
-
-        if (!strcmp(entry->d_name, ".") || !strcmp(entry->d_name, ".."))
-            continue;
-
-        if (nats_asprintf(&fullPath, "%s/%s", path, entry->d_name) < 0)
-            abort();
-
-        memset(&statEntry, 0, sizeof(struct stat));
-        stat(fullPath, &statEntry);
-
-        if (S_ISDIR(statEntry.st_mode) != 0)
-            rmtree(fullPath);
-        else
-            unlink(fullPath);
-
-        free(fullPath);
-    }
-
-    closedir(dir);
-    rmdir(path);
-#endif
-}
-
-static void
-_makeUniqueDir(char *buf, int bufLen, const char *path)
-{
-    int n;
-
-    if ((int) strlen(path) + 1 + NUID_BUFFER_LEN + 1 > bufLen)
-        abort();
-
-    n = snprintf(buf, bufLen, "%s", path);
-    natsNUID_Next(buf+n, NUID_BUFFER_LEN+1);
-    buf[n+NUID_BUFFER_LEN+1] = '\0';
-}
-
-static void
 _createConfFile(char *buf, int bufLen, const char *content)
 {
     FILE *f = NULL;
@@ -43467,9 +43356,11 @@ int main(int argc, char **argv)
     if (failed)
     {
         printf("*** TEST FAILED ***\n");
+        fflush(stdout);
         return 1;
     }
 
     printf("ALL PASSED\n");
+    fflush(stdout);
     return 0;
 }
